Fix load on Android 11

+++ app/src/main/java/com/o4x/musical/model/Genre.java
@@ -8,19 +8,19 @@ import java.util.List;
 
 public class Genre implements Parcelable {
 
-    public final int id;
+    public final long id;
     public final String name;
     public final int songCount;
     public final List<Song> songs;
 
-    public Genre(final int id, final String name, final int songCount, final List<Song> songs) {
+    public Genre(final long id, final String name, final int songCount, final List<Song> songs) {
         this.id = id;
         this.name = name;
         this.songCount = songCount;
         this.songs = songs;
     }
 
-    public Genre(final int id, final String name, final int songCount) {
+    public Genre(final long id, final String name, final int songCount) {
         this.id = id;
         this.name = name;
         this.songCount = songCount;
@@ -42,11 +42,11 @@ public class Genre implements Parcelable {
 
     @Override
     public int hashCode() {
-        int result = id;
+        long result = id;
         result = 31 * result + name.hashCode();
         result = 31 * result + songCount;
         result = 31 * result + (songs != null ? songs.hashCode() : 0);
-        return result;
+        return (int)result;
     }
 
     @Override
@@ -66,14 +66,14 @@ public class Genre implements Parcelable {
 
     @Override
     public void writeToParcel(Parcel dest, int flags) {
-        dest.writeInt(this.id);
+        dest.writeLong(this.id);
         dest.writeString(this.name);
         dest.writeInt(this.songCount);
         dest.writeTypedList(this.songs);
     }
 
     protected Genre(Parcel in) {
-        this.id = in.readInt();
+        this.id = in.readLong();
         this.name = in.readString();
         this.songCount = in.readInt();
         this.songs = in.createTypedArrayList(Song.CREATOR);
