+++ app/src/main/java/com/o4x/musical/ui/activities/details/AlbumDetailActivity.java
@@ -427,7 +427,7 @@ public class AlbumDetailActivity extends AbsMusicPanelActivity implements Palett
 
     @Override
     public Loader<Album> onCreateLoader(int id, Bundle args) {
-        return new AsyncAlbumLoader(this, args.getInt(EXTRA_ALBUM_ID));
+        return new AsyncAlbumLoader(this, args.getLong(EXTRA_ALBUM_ID));
     }
 
     @Override
@@ -442,9 +442,9 @@ public class AlbumDetailActivity extends AbsMusicPanelActivity implements Palett
     }
 
     private static class AsyncAlbumLoader extends WrappedAsyncTaskLoader<Album> {
-        private final int albumId;
+        private final long albumId;
 
-        public AsyncAlbumLoader(Context context, int albumId) {
+        public AsyncAlbumLoader(Context context, long albumId) {
             super(context);
             this.albumId = albumId;
         }
